% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rimas.R
\name{read_rimas}
\alias{read_rimas}
\title{Function to process RIMAS data}
\usage{
read_rimas(file)
}
\arguments{
\item{file}{String. A path to the file downloaded from the RIMAS website.}
}
\value{
\code{read_rimas} returns an \code{data.frame}
}
\description{
Clean and transform data of the Groundwater Monitoring Network (RIMAS) from
Geological Survey of Brazil (CPRM).
Data from groundwater monitoring wells can be found at
\link{http://rimasweb.cprm.gov.br/layout/}.
}
\seealso{
\code{\link{read_funceme}}
}
\author{
Rubens da Cunha Junior
}
